PPodEntity is no longer an entity and just a mappedsuperclass because Attachment no longer needed it as a persisted backref. Now using single table per class hierarcy. This gave us a good upload speed improvement.
Only Character has an attachment. This gives us a great download speed improvement. For some reason, it looks like having the cells have attachments, even with a mappedsuperclass PPodEntity, is not possible because of the slow down it causes. One possible other solution is to keep a two dimensional array of attachemnts in the matrix.